refactor: move correlatedObject to its own file
no changes except package naming
